requirements-dev.txt
-r requirements.txt
django-silk==5.3.2
django-extensions==3.2.3
factory_boy==3.3.3